full app migration, fix modal
