+++ packages/app/src/components/Root/index.ts
@@ -1,0 +1,1 @@
+export { navModule } from './navModule';
