+++ packages/app/src/components/Root/navModule.tsx
@@ -1,0 +1,82 @@
+import { createFrontendModule } from '@backstage/frontend-plugin-api';
+import { NavContentBlueprint } from '@backstage/plugin-app-react';
+import {
+  Sidebar,
+  SidebarDivider,
+  SidebarGroup,
+  SidebarItem,
+  SidebarScrollWrapper,
+  sidebarConfig,
+  useSidebarOpenState,
+  Link,
+} from '@backstage/core-components';
+import { SidebarSearchModal } from '@backstage/plugin-search';
+import { makeStyles } from '@material-ui/core';
+import MenuIcon from '@material-ui/icons/Menu';
+import SearchIcon from '@material-ui/icons/Search';
+import FeedbackIcon from '@material-ui/icons/Feedback';
+import { OpenFeedbackModal } from '@parsifal-m/backstage-plugin-open-feedback';
+import LogoFull from './LogoFull';
+import LogoIcon from './LogoIcon';
+
+const useSidebarLogoStyles = makeStyles({
+  root: {
+    width: sidebarConfig.drawerWidthClosed,
+    height: 3 * sidebarConfig.logoHeight,
+    display: 'flex',
+    flexFlow: 'row nowrap',
+    alignItems: 'center',
+    marginBottom: -14,
+  },
+  link: {
+    width: sidebarConfig.drawerWidthClosed,
+    marginLeft: 24,
+  },
+});
+
+const SidebarLogo = () => {
+  const classes = useSidebarLogoStyles();
+  const { isOpen } = useSidebarOpenState();
+  return (
+    <div className={classes.root}>
+      <Link to="/" underline="none" className={classes.link} aria-label="Home">
+        {isOpen ? <LogoFull /> : <LogoIcon />}
+      </Link>
+    </div>
+  );
+};
+
+const navContent = NavContentBlueprint.make({
+  params: {
+    component: ({ navItems }) => {
+      const nav = navItems.withComponent(item => (
+        <SidebarItem icon={() => item.icon} to={item.href} text={item.title} />
+      ));
+      return (
+        <Sidebar>
+          <SidebarLogo />
+          <SidebarGroup label="Search" icon={<SearchIcon />} to="/search">
+            <SidebarSearchModal />
+            <OpenFeedbackModal
+              floating
+              title="OpenFeedback"
+              rating={2}
+              disableAnonymous={false}
+              icon={FeedbackIcon}
+              style={{ position: 'fixed', bottom: 20, right: 20 }}
+            />
+          </SidebarGroup>
+          <SidebarDivider />
+          <SidebarGroup label="Menu" icon={<MenuIcon />}>
+            <SidebarScrollWrapper>{nav.rest()}</SidebarScrollWrapper>
+          </SidebarGroup>
+        </Sidebar>
+      );
+    },
+  },
+});
+
+export const navModule = createFrontendModule({
+  pluginId: 'app',
+  extensions: [navContent],
+});
