+++ plugins/open-feedback/package.json
@@ -73,7 +73,7 @@
     "@backstage/cli": "^0.36.3",
     "@backstage/core-app-api": "^1.20.2",
     "@backstage/dev-utils": "^1.1.24",
-    "@backstage/frontend-defaults": "^0.3.3",
+    "@backstage/frontend-defaults": "^0.5.3",
     "@backstage/test-utils": "^1.7.19",
     "@testing-library/jest-dom": "^6.0.0",
     "@testing-library/react": "^14.0.0",
